Extract JsonLogProcessor module
