+++ index.js
@@ -6,9 +6,9 @@ const { ERROR_LIKE_KEYS, MESSAGE_KEY, TIMESTAMP_KEY } = require('./lib/constants
 
 const {
   defaultLogParsingSequence,
-  createLogProcessor,
-  JsonLogProcessor
+  createLogProcessor
 } = require('./lib/log-processors')
+const JsonLogProcessor = require('./lib/processors/JsonLogProcessor')
 const { buildLine } = require('./lib/line-builders')
 
 const defaultOptions = {
