+++ lib/log-parsers.js
@@ -2,31 +2,6 @@
 
 const jmespath = require('jmespath')
 
-const {
-  isObject
-} = require('./utils')
-
-const bourne = require('@hapi/bourne')
-const jsonParser = input => {
-  try {
-    return { value: bourne.parse(input, { protoAction: 'remove' }) }
-  } catch (err) {
-    return { err }
-  }
-}
-
-function json (input, { EOL }) {
-  if (!isObject(input)) {
-    const parsed = jsonParser(input)
-    if (parsed.err) {
-      // pass through
-      return { output: input + EOL, done: true }
-    } else {
-      return { output: parsed.value }
-    }
-  } else return { output: input }
-}
-
 function primitives (input) {
   // Short-circuit for spec allowed primitive values.
   if ([null, true, false].includes(input) || Number.isFinite(input)) {
@@ -60,7 +35,6 @@ function ignore (input, { ignore }) {
 }
 
 const builtInLogParsers = {
-  json,
   primitives,
   search,
   ignore
