+++ lib/log-processors.js
@@ -2,8 +2,9 @@ const { builtInLogParsers } = require('./log-parsers')
 const { prettifyLevel, prettifyMessage, prettifyMetadata, prettifyTime } = require('./utils')
 const { builtInLineBuilders } = require('./line-builders')
 const LogProcessor = require('./LogProcessor')
+const JsonLogProcessor = require('./processors/JsonLogProcessor')
 
-const { json, primitives, search, ignore } = builtInLogParsers
+const { primitives, search, ignore } = builtInLogParsers
 const {
   appendLevel,
   appendMessage,
@@ -26,12 +27,6 @@ function createLogProcessor (definition) {
     return new builtInLogProcessors[definition]()
   } else {
     return definition
-  }
-}
-
-class JsonLogProcessor extends LogProcessor {
-  parse (input, context) {
-    return json(input, context)
   }
 }
 
@@ -141,6 +136,5 @@ const defaultLogParsingSequence = [
 module.exports = {
   builtInLogProcessors,
   defaultLogParsingSequence,
-  createLogProcessor,
-  JsonLogProcessor
+  createLogProcessor
 }
